Minor modifications on LongRange and ChunkedList
The ChunkedList class had an issue with the way the keys were handled.
It now uses the normal LongRange key order.
Also, the get/set/contains methods needed to be adapted to take into
account multiple cases, including whether the index is on the left side
of a chunk or inside of one. There might be a simple implementation for
these methods, but for now it seems to work.