Converted the test routines to use the Junit framework with our custom MpiRunner.
All the distributed tests are Integration Tests (class name starts with IT). They will be run by using `mvn verify`. A number of changes were made to the pom.xml, in particular we now use the latest version of the surefire/failsafe plugin (version 3.0.0-M5) to avoid getting errors when these test programs make outputs to STDOUT or STDERR.