Implemented the DistCol GLB operations using a LongRange as input.
This changes from the Long which was previously used for all
GlbOperations. It allows us to move the loop on the elements of the
DistCol from the DistColGlbTask#Assignment to the GlbOperation#action.
As a result, the map operation can now be implemented without checking
if a Chunk is present in the result DistCol for every element
considered.