Re-wrote how chunks are split in DistCol
In the previous implementation, the chunk which was being split was removed first. Then the multiple chunks that are placed into the ChunkedList as a replacement are inserted. This posed a problem for the GLB. When an assignment is split into multiple ones, the underlying chunk is not. If one of these assignment is then sent as part of a lifeline answer, there is a short period of time during the chunk splitting operation when a neighboring assignment could try (and fail) to access the data, causing some GLB runtime errors.

This bug is now fixed by inserting the smaller chunks into the ChunkedList BEFORE removing the original (larger) chunk. As a result, there is a transitional stage where a ChunkedList contains overlapping Chunks (which is disallowed through the external ChunkedList / DistCol API).